更改包名、应用名、主页面名
- 更改包的名字为com.google.ar.core.ARPositioning.kotlin.Transfer
- 更改应用id为com.Alex.ar.core.kotlin.ARPositioning
- 更改主页面的类名为mainActivity
- 更改显示类名为ArView
- 更改渲染类名为ArRenderer
