+++ app/build.gradle
@@ -4,14 +4,14 @@ apply plugin: 'kotlin-android'
 android {
     compileSdkVersion 31
     defaultConfig {
-        applicationId "com.google.ar.core.examples.kotlin.helloar"
+        applicationId "com.Alex.ar.core.kotlin.ARPositioning"
 
         // AR Optional apps must declare minSdkVersion >= 14.
         // AR Required apps must declare minSdkVersion >= 24.
         minSdkVersion 24
         targetSdkVersion 31
         versionCode 1
-        versionName '1.0'
+        versionName '2.1'
     }
     compileOptions {
         sourceCompatibility JavaVersion.VERSION_1_8
@@ -29,6 +29,14 @@ android {
     viewBinding {
         enabled = true
     }
+    android.applicationVariants.all { variant ->
+        variant.outputs.all {
+            outputFileName = "ARPositioning_${buildType.name}_v${versionName}_${generateTime()}.apk"
+        }
+    }
+}
+def generateTime() {
+    return new Date().format("MMdd_HHmm")
 }
 
 dependencies {
