+++ app/src/main/java/com/google/ar/core/ARPositioning/java/common/helpers/DepthSettings.java
@@ -1,0 +1,88 @@
+/*
+ * Copyright 2020 Google LLC
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *   http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package com.google.ar.core.ARPositioning.java.common.helpers;
+
+import android.content.Context;
+import android.content.SharedPreferences;
+
+/** Manages the Occlusion option setting and shared preferences.
+ * 管理遮挡选项设置和共享首选项。*/
+public class DepthSettings {
+  public static final String SHARED_PREFERENCES_ID = "SHARED_PREFERENCES_OCCLUSION_OPTIONS";
+  public static final String SHARED_PREFERENCES_SHOW_DEPTH_ENABLE_DIALOG_OOBE =
+      "show_depth_enable_dialog_oobe";
+  public static final String SHARED_PREFERENCES_USE_DEPTH_FOR_OCCLUSION = "use_depth_for_occlusion";
+
+  // Current depth-based settings used by the app.
+  private boolean depthColorVisualizationEnabled = false;
+  private boolean useDepthForOcclusion = false;
+  private SharedPreferences sharedPreferences;
+
+  /** Initializes the current settings based on when the app was last used.
+   * 根据上次使用应用程序的设置初始化当前设置*/
+  public void onCreate(Context context) {
+    sharedPreferences = context.getSharedPreferences(SHARED_PREFERENCES_ID, Context.MODE_PRIVATE);
+    useDepthForOcclusion =
+        sharedPreferences.getBoolean(SHARED_PREFERENCES_USE_DEPTH_FOR_OCCLUSION, false);
+  }
+
+  /** Retrieves whether depth-based occlusion is enabled.
+   * 检索是否启用了基于深度的遮挡*/
+  public boolean useDepthForOcclusion() {
+    return useDepthForOcclusion;
+  }
+
+  public void setUseDepthForOcclusion(boolean enable) {
+    if (enable == useDepthForOcclusion) {
+      return; // No change. 如果设置没有变化回退即可
+    }
+
+    // Updates the stored default settings. 设置变化了就使用enable来更新
+    useDepthForOcclusion = enable;
+    SharedPreferences.Editor editor = sharedPreferences.edit();
+    editor.putBoolean(SHARED_PREFERENCES_USE_DEPTH_FOR_OCCLUSION, useDepthForOcclusion);
+    editor.apply();
+  }
+
+  /** Retrieves whether to render the depth map visualization instead of the camera feed.
+   * 检索是否渲染深度图可视化而不是相机源。*/
+  public boolean depthColorVisualizationEnabled() {
+    return depthColorVisualizationEnabled;
+  }
+
+  public void setDepthColorVisualizationEnabled(boolean depthColorVisualizationEnabled) {
+    this.depthColorVisualizationEnabled = depthColorVisualizationEnabled;
+  }
+
+  /** Determines if the initial prompt to use depth-based occlusion should be shown.
+   * 确定是否应显示使用基于深度的遮挡的初始提示。*/
+  public boolean shouldShowDepthEnableDialog() {
+    // Checks if this dialog has been called before on this device. 检查之前是否在此设备上调用过此对话框。
+    boolean showDialog =
+        sharedPreferences.getBoolean(SHARED_PREFERENCES_SHOW_DEPTH_ENABLE_DIALOG_OOBE, true);
+
+    if (showDialog) {
+      // Only ever shows the dialog on the first time.  If the user wants to adjust these settings
+      // 只在第一次显示对话框。如果用户想要再次调整这些设置，他们可以使用齿轮图标来调用设置菜单对话框。
+      // again, they can use the gear icon to invoke the settings menu dialog.
+      SharedPreferences.Editor editor = sharedPreferences.edit();
+      editor.putBoolean(SHARED_PREFERENCES_SHOW_DEPTH_ENABLE_DIALOG_OOBE, false);
+      editor.apply();
+    }
+
+    return showDialog;
+  }
+}
