+++ app/src/main/java/com/google/ar/core/ARPositioning/kotlin/Temp.java
@@ -1,0 +1,28 @@
+package com.google.ar.core.ARPositioning.kotlin;
+
+import com.google.ar.core.ARPositioning.kotlin.Transfer.R;
+
+public class Temp {
+    public enum AppState{
+        Idle,
+        Recording
+    };
+    private AppState appState = AppState.Idle;
+
+
+//    private void updateRecordButton() {
+//        View buttonView = findViewById(R.id.record_button);
+//        Button button = (Button) buttonView;
+//
+//        switch (appState) {
+//            case Idle:
+//                button.setText("Record");
+//                break;
+//            case Recording:
+//                button.setText("Stop");
+//                break;
+//        }
+//    }
+
+
+}
