+++ app/src/main/java/com/google/ar/core/ARPositioning/kotlin/temp.kt
@@ -1,0 +1,11 @@
+package com.google.ar.core.ARPositioning.kotlin
+
+
+//private fun updateRecordButton() {
+//    val buttonView: View = findViewById(R.id.record_button)
+//    val button = buttonView as Button
+//    when (appState) {
+//        Temp.AppState.Idle -> button.text = "Record"
+//        Temp.AppState.Recording -> button.text = "Stop"
+//    }
+//}
