--- app/src/main/java/com/google/ar/core/examples/kotlin/temp.kt
@@ -1,14 +1,0 @@
-package com.google.ar.core.examples.kotlin
-
-import com.google.ar.core.*
-
-
-
-//private fun updateRecordButton() {
-//    val buttonView: View = findViewById(R.id.record_button)
-//    val button = buttonView as Button
-//    when (appState) {
-//        Temp.AppState.Idle -> button.text = "Record"
-//        Temp.AppState.Recording -> button.text = "Stop"
-//    }
-//}
